requirements.txt
gensim
scikit-learn
tqdm
python-dotenv
numpy<2
scipy